backend/auth_service/requirements.txt
fastapi==0.68.0
pyjwt==2.3.0
uvicorn==0.15.0
python-multipart==0.0.5
pymongo==4.0.1